requirements.txt
asgiref==3.8.1
colorama==0.4.6
decorator==5.2.1
Django==5.1.7
django-crispy-forms==2.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.1.2
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
numpy==2.2.4
pillow==10.4.0
proglog==0.1.10
pyarrow==19.0.1
pylance==0.24.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.1

boto3==1.38.16
django-storages==1.14.6
ffmpeg==1.4
#opencv-python==4.11.0.86
django-ajax-datatable==4.5.0
celery==5.5.2

